backend/pyproject.toml
[tool.poetry]
name = "dnd-portfolio-project"
version = "0.1.0"
description = ""
authors = ["AliiMustafa <aliimustafazada@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.12"}
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["all"], version = "^2.0.40"}
alembic = "^1.15.2"
aiomysql = "^0.2.0"
jos = "^0.1.2"
jose = "^1.0.0"
python-jose = "^3.4.0"
bcrypt = "^4.3.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.10.1"
pytest = "^8.3.5"
pymysql = "^1.1.1"
openai = "^1.69.0"
cryptography = "^44.0.2"
python-dotenv = "^1.1.0"
redis = "^5.2.1"
aiosmtplib = "^4.0.0"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
